quizz_generator/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
attrs==24.2.0
azure-core==1.31.0
azure-storage-blob==12.22.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dill==0.3.8
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
isodate==0.6.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.1.0
openai==0.28.0
platformdirs==4.3.3
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pylint==3.2.7
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
importlib_metadata